Pulling lots of recent changes from Lucas and Melissa

+++ run_imi.sh
@@ -36,6 +36,16 @@ fi
 source src/utilities/parse_yaml.sh
 eval $(parse_yaml ${ConfigFile})
 
+if ! "$isAWS"; then
+    # Activate Conda environment
+    printf "\nActivating conda environment: ${CondaEnv}\n"
+    eval "$(conda shell.bash hook)"
+    conda activate $CondaEnv
+fi
+
+# Check all necessary config variables are present
+python src/utilities/sanitize_input_yaml.py $ConfigFile || imi_failed
+
 # Set path to IMI runs
 RunDirs="${OutputPath}/${RunName}"
 
@@ -98,7 +108,9 @@ if "$isAWS"; then
     printf "\nFinished TROPOMI download\n"
 else
     # use existing tropomi data and create a symlink to it
-    ln -s $DataPathTROPOMI $tropomiCache
+    if [[ ! -L $tropomiCache ]]; then
+	ln -s $DataPathTROPOMI $tropomiCache
+    fi
 fi
 
 ##=======================================================================
@@ -241,6 +253,7 @@ if "$DoPosterior"; then
 
     # Build directory for hourly posterior GEOS-Chem output data
     mkdir -p data_converted_posterior
+    mkdir -p data_visualization_posterior
     mkdir -p data_geoschem_posterior
     GCsourcepth="${PosteriorRunDir}/OutputDir"
     GCDir="./data_geoschem_posterior"
@@ -290,5 +303,9 @@ printf "\n Posterior : $(( $posterior_end - $posterior_start ))\n\n"
 if [[ -f ${InversionPath}/imi_output.log ]]; then
     cp "${InversionPath}/imi_output.log" "${RunDirs}/imi_output.log"
 fi
+
+# copy config file to run directory
+cd $InversionPath
+cp $ConfigFile "${RunDirs}/config_${RunName}.yml"
 
 exit 0
